refactor: move collection_from_items script to topo-imagery-raster TDE-2050
