+++ packages/topo-imagery-raster/test/conftest.py
@@ -1,0 +1,1 @@
+from topo_imagery_stac.testing.helpers import fake_collection_context  # pylint: disable=unused-import
